cybertwin-backend/requirements.txt
# ── Core Framework ────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0

# ── Pydantic / Config ─────────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

# ── Security (Phase 3 stubs) ──────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ── Phase 3 AI/NLP (pre-installed, wired in Phase 3) ─────────────────────────
openai>=1.35.0

# ── Phase 2 Threat Intel – MITRE ATT&CK STIX 2.0 ─────────────────────────────
mitreattack-python>=3.0.6
